requirements.txt
numpy==1.19.2
pandas==1.1.3
progressbar33==2.4
PyYAML==5.4.1
